requirements.txt
torch
numpy
python-chess
stockfish
flask
requests